fix(simulator,ui): correct rest-priority, float break counter, and midnight pill
Engine (engine.py / models.py / rules.py):
- Mandatory rest now unconditionally wins over a coincident 30-min break in
  the post-split event-insertion chain.  Previously, when both break_hit and
  rest_hit fired simultaneously the engine inserted an unnecessary 30-min
  break before the 10h rest; the 10h rest resets driving_minutes_since_break
  anyway, so the break was redundant and illegal (on-duty work performed after
  the driving cap was reached).  This also eliminated the spurious "10h 30m"
  merged rest run that made the boundary pill and DOT 11h check appear to
  disagree.
- driving_minutes_since_break changed from int to float (models.py) and all
  int() truncation casts removed from the two arithmetic sites in engine.py,
  so sub-minute segments accumulate without rounding loss.
- Corresponding 0 → 0.0 literals updated in rules.py for consistency.

Frontend (TimelineView.tsx):
- Replaced the per-day, minute-of-day–scoped compliance logic with a single
  computeCompliance pass that walks all events as one continuous timestamp
  stream.  The DOT 11h check now accumulates driving since the last true
  ≥10h rest run, never resetting at calendar midnight.  Both the DOT 11h
  check and the between-day boundary pill read from the same detectRestRuns
  output, so they cannot disagree.
- Between-day rest pill is suppressed when detected boundary rest is 0 min
  (driver legally crossing midnight within one duty window); green ≥10h and
  34h restart pills still render normally.

Tests:
- Added test_fmcsa_09_chicago_mustang_norwalk_cycle69_no_midnight_chain to
  test_fmcsa_scenarios.py: exact bug-report input (Chicago → Mustang OK →
  Norwalk CA, cycle 69h) with three invariant checks — 11h cap per window,
  every mandatory rest ≥10h, per-day peak matches cross-day accumulation.
- Added test_continuous_rest_11h.py: two tests that mirror frontend semantics
  (detectRestRuns + drivingMaxByDay) and assert pill/check agreement for the
  Chicago → Denver → LA cycle-65 scenario.
- All 37 tests pass.

+++ backend/simulator/models.py
@@ -67,7 +67,10 @@ class SimulationState:
     window_start: datetime
     window_end: datetime
     driving_hours_today: float = 0.0
-    driving_minutes_since_break: int = 0
+    # Minutes driven since the last qualifying break. Kept as float so small
+    # clamped segments (e.g. after an 11h or 14h cap split) do not lose
+    # sub-minute time and silently delay the mandatory 30-min break.
+    driving_minutes_since_break: float = 0.0
     cycle_hours_used: float = 0.0
     miles_since_fuel: float = 0.0
     restart_used: bool = False
